requirements.txt
requests
beautifulsoup4
html2text
html5lib
wxPython
